Don't re-calculate obviously static filenames
Filenames that are enclosed in [] are entirely static and don't need
to be re-interpretted for every log message.

In cases where simple data is being logged to a static file, the time
formatting used for the filename can account for 50% or more of the
time spent in a log() call.